Move country fetching
