+++ src/frontend/app/(pages)/country/page.tsx
@@ -1,4 +1,4 @@
-import { fetchCountryList } from '@/lib/api/dataFetching';
+import { fetchCountryList } from '@/lib/api/country';
 import { paginate } from '@/lib/helpers/pagination';
 import { Pagination } from '@/components/custom/pagination';
 import { PageTitle } from '@/components/custom/page-title';
